backend/requirements/local.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --output-file /local/local.txt /local/local.in
argon2-cffi==25.1.0
    # via -r local/base.txt
argon2-cffi-bindings==21.2.0
    # via
    #   -r local/base.txt
    #   argon2-cffi
asgiref==3.9.1
    # via
    #   -r local/base.txt
    #   daphne
    #   django
    #   django-allauth
    #   django-cors-headers
    #   strawberry-graphql-django
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   -r local/base.txt
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via
    #   -r local/base.txt
    #   daphne
automat==25.4.16
    # via
    #   -r local/base.txt
    #   twisted
cffi==1.17.1
    # via
    #   -r local/base.txt
    #   argon2-cffi-bindings
    #   cryptography
constantly==23.10.4
    # via
    #   -r local/base.txt
    #   twisted
crispy-bootstrap5==2025.6
    # via -r local/base.txt
cryptography==45.0.5
    # via
    #   -r local/base.txt
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.2.1
    # via -r local/base.txt
debugpy==1.8.15
    # via -r local/local.in
decorator==5.2.1
    # via
    #   ipdb
    #   ipython
django==5.2.4
    # via
    #   -r local/base.txt
    #   crispy-bootstrap5
    #   django-allauth
    #   django-cors-headers
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-model-utils
    #   django-redis
    #   strawberry-graphql-django
django-allauth==65.10.0
    # via -r local/base.txt
django-cors-headers==4.7.0
    # via -r local/base.txt
django-crispy-forms==2.4
    # via
    #   -r local/base.txt
    #   crispy-bootstrap5
django-debug-toolbar==5.2.0
    # via -r local/local.in
django-environ==0.12.0
    # via -r local/base.txt
django-extensions==4.1
    # via -r local/local.in
django-model-utils==5.0.0
    # via -r local/base.txt
django-redis==6.0.0
    # via -r local/base.txt
executing==2.2.0
    # via stack-data
graphql-core==3.2.6
    # via
    #   -r local/base.txt
    #   strawberry-graphql
hyperlink==21.0.0
    # via
    #   -r local/base.txt
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   -r local/base.txt
    #   hyperlink
    #   twisted
incremental==24.7.2
    # via
    #   -r local/base.txt
    #   twisted
ipdb==0.13.13
    # via -r local/local.in
ipython==9.4.0
    # via ipdb
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
markupsafe==3.0.2
    # via werkzeug
matplotlib-inline==0.1.7
    # via ipython
packaging==25.0
    # via
    #   -r local/base.txt
    #   strawberry-graphql
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.3.0
    # via -r local/base.txt
prompt-toolkit==3.0.51
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   -r local/base.txt
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via
    #   -r local/base.txt
    #   service-identity
pycparser==2.22
    # via
    #   -r local/base.txt
    #   cffi
pydevd-pycharm==243.26053.29
    # via -r local/local.in
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pyopenssl==25.1.0
    # via
    #   -r local/base.txt
    #   twisted
python-dateutil==2.9.0.post0
    # via
    #   -r local/base.txt
    #   strawberry-graphql
redis==6.2.0
    # via
    #   -r local/base.txt
    #   django-redis
service-identity==24.2.0
    # via
    #   -r local/base.txt
    #   twisted
setuptools==80.9.0
    # via
    #   -r local/base.txt
    #   autobahn
    #   incremental
    #   zope-interface
six==1.17.0
    # via
    #   -r local/base.txt
    #   python-dateutil
sqlparse==0.5.3
    # via
    #   -r local/base.txt
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
strawberry-graphql==0.278.0
    # via
    #   -r local/base.txt
    #   strawberry-graphql-django
strawberry-graphql-django==0.65.0
    # via -r local/base.txt
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
twisted==25.5.0
    # via
    #   -r local/base.txt
    #   daphne
txaio==25.6.1
    # via
    #   -r local/base.txt
    #   autobahn
typing-extensions==4.14.1
    # via
    #   -r local/base.txt
    #   pyopenssl
    #   strawberry-graphql
    #   twisted
urllib3==2.5.0
    # via -r local/base.txt
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via -r local/local.in
zope-interface==7.2
    # via
    #   -r local/base.txt
    #   twisted
